Move parser into AST module
